==> src/chambers11/2_build_corpus.py <==
"""
In their original publication, the authors expand the MUC corpus with 3.5 million additional newswire articles
before inducing the semantic roles (template slots). I will not collect more data, but this code will instead 
use the infromation from 1_template_topics.py to assign documents to clusters. 
This assignment to clusters is essential for 3_template_slots.py
"""
